+++ src/monaco/mc_enums.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 from enum import Enum
 
+
 class SampleMethod(str, Enum):
     """
     Enum for the possible Monte Carlo sampling methods.